update BankCollector test
