+++ sample/__tests__/_/giver.js
@@ -6,6 +6,17 @@ dotenv.config();
 const nodeSe = !!process.env.USE_NODE_SE
     && process.env.USE_NODE_SE.toLowerCase() !== 'false'
     && process.env.USE_NODE_SE !== '0';
+
+function loadPackage(name) {
+    const base = path.resolve(process.cwd(), '__tests__', 'contracts');
+    const abi = path.resolve(base, `${name}.abi.json`);
+    const tvc = path.resolve(base, `${name}.tvc`);
+    return {
+        abi: JSON.parse(fs.readFileSync(abi, 'utf8')),
+        imageBase64: fs.readFileSync(tvc).toString('base64'),
+    }
+}
+
 async function readGiverKeys(client) {
     try {
         let keysPath = path.resolve(os.homedir(), 'giverKeys.json');
@@ -113,7 +124,7 @@ const GiverWalletPackage = {
     },
     imageBase64: 'te6ccgECGgEAA9sAAgE0BgEBAcACAgPPIAUDAQHeBAAD0CAAQdgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAIm/wD0pCAiwAGS9KDhiu1TWDD0oQkHAQr0pCD0oQgAAAIBIAwKAfz/fyHtRNAg10nCAZ/T/9MA9AX4an/4Yfhm+GKOG/QFbfhqcAGAQPQO8r3XC//4YnD4Y3D4Zn/4YeLTAAGOEoECANcYIPkBWPhCIPhl+RDyqN4j+EL4RSBukjBw3rry4GUh0z/THzQx+CMhAb7yuSH5ACD4SoEBAPQOIJEx3rMLAE7y4Gb4ACH4SiIBVQHIyz9ZgQEA9EP4aiMEXwTTHwHwAfhHbpLyPN4CASASDQIBWBEOAQm46Jj8UA8B/vhBbo4S7UTQ0//TAPQF+Gp/+GH4Zvhi3tFwbW8C+EqBAQD0hpUB1ws/f5NwcHDikSCONyMjI28CbyLIIs8L/yHPCz8xMQFvIiGkA1mAIPRDbwI0IvhKgQEA9HyVAdcLP3+TcHBw4gI1MzHoXwPIghB3RMfighCAAAAAsc8LHyEQAKJvIgLLH/QAyIJYYAAAAAAAAAAAAAAAAM8LZoEDmCLPMQG5lnHPQCHPF5Vxz0EhzeIgyXH7AFswwP+OEvhCyMv/+EbPCwD4SgH0AMntVN5/+GcAxbkWq+f/CC3Rxt2omgQa6ThAM/p/+mAegL8NT/8MPwzfDFHDfoCtvw1OADAIHoHeV7rhf/8MTh8Mbh8Mz/8MPFvfCNJeRnJuPwzcXwAaPwhZGX//CNnhYB8JQD6AGT2qj/8M8AIBIBUTAde7Fe+TX4QW6OEu1E0NP/0wD0Bfhqf/hh+Gb4Yt76QNcNf5XU0dDTf9/XDACV1NHQ0gDf0SIiInPIcc8LASLPCgBzz0AkzxYj+gKAac9Acs9AIMki+wBfBfhKgQEA9IaVAdcLP3+TcHBw4pEggUAJKOLfgjIgG7n/hKIwEhAYEBAPRbMDH4at4i+EqBAQD0fJUB1ws/f5NwcHDiAjUzMehfA18D+ELIy//4Rs8LAPhKAfQAye1Uf/hnAgEgFxYAx7jkYYdfCC3Rwl2omhp/+mAegL8NT/8MPwzfDFvamj8IXwikDdJGDhvXXlwMvwAfCFkZf/8I2eFgHwlAPoAZPaqfAeQfYIQaHaPdqn4ARh8IWRl//wjZ4WAfCUA+gBk9qo//DPACAtoZGAAtr4QsjL//hGzwsA+EoB9ADJ7VT4D/IAgAdacCHHAJ0i0HPXIdcLAMABkJDi4CHXDR+S8jzhUxHAAJDgwQMighD////9vLGS8jzgAfAB+EdukvI83o',
 };
-const giverRequestAmount = 2_000_000_000;
+const giverRequestAmount = 5_000_000_000;
 
 async function check_giver(client) {
 
@@ -133,7 +144,7 @@ async function check_giver(client) {
     }
 }
 
-async function get_grams_from_giver(client, account,amount) {
+async function get_grams_from_giver(client, account, amount) {
     const { contracts, queries } = client;
 
     console.time(`Get grams from giver to ${account}`);
@@ -166,24 +177,20 @@ async function get_grams_from_giver(client, account,amount) {
     const result = await contracts.run(params);
     for (const msg in (result.transaction.out_messages || [])) {
         if (msg.msg_type === 0) {
-           // config.log(`Giver. Wait for ${msg.id || "Empty ID"}`);
+            // config.log(`Giver. Wait for ${msg.id || "Empty ID"}`);
             await queries.transactions.waitFor(
                 {
                     in_msg: { eq: msg.id },
-                    status: { eq: 5},
+                    status: { eq: 5 },
                 },
                 'lt'
             );
         }
     }
-
-   // config.log("Giver. End");
     console.timeEnd(`Get grams from giver to ${account}`);
-
 }
 
 async function deploy_with_giver(client, params) {
-
     const { contracts } = client;
     const address = (await contracts.getDeployData({
         ...params.package,
@@ -197,24 +204,8 @@ async function deploy_with_giver(client, params) {
     return contracts.deploy(params);
 }
 
-
-function get_giver_address() {
-    return nodeSe ? nodeSeGiverAddress : giverWalletAddressHex;
-}
-
-function add_deployed_contract(key, address, abi) {
-    tests.deployedContracts.push({
-        key,
-        address,
-        abi,
-        giverAddress: nodeSe ? nodeSeGiverAddress : giverWalletAddressHex,
-    });
-}
-
 module.exports = {
-    add_deployed_contract,
-    get_grams_from_giver,
+    loadPackage,
     deploy_with_giver,
-    check_giver,
     readGiverKeys
 }
